requirements.txt
azure-storage-blob
azure-storage-queue
python-dotenv
